refactors application and adds readme
