+++ .dockerignore
@@ -1,0 +1,9 @@
+Dockerfile
+.dockerignore
+node_modules
+npm-debug.log
+.git
+.DS_Store
+.nyc_output
+.env
+.data
