+++ Dockerfile
@@ -1,0 +1,12 @@
+FROM node:14.16.1-alpine3.10 as build-step
+RUN mkdir /app
+WORKDIR /app
+COPY package.json /app
+RUN npm install
+COPY . /app
+RUN npm run build
+
+FROM nginx
+COPY --from=build-step /app/dist /usr/share/nginx/html
+COPY nginx/nginx.conf /etc/nginx/conf.d/default.conf
+CMD ["nginx", "-g", "daemon off;"]
