+++ nginx/nginx.conf
@@ -1,0 +1,18 @@
+server {
+        listen 80;
+        listen [::]:80;
+
+        server_name modelling.com  www.modelling.com;
+
+          location / {
+                root /usr/share/nginx/html/;
+                index index.html;
+                try_files $uri $uri/ /index.html;
+
+    }
+
+        location /api {
+                proxy_pass http://127.0.0.1:4000;
+                proxy_set_header Host $host;
+        }
+}
