+++ package.json
@@ -5,7 +5,8 @@
   "main": "index.js",
   "scripts": {
     "lint": "eslint --init",
-    "start": "npm run build && node server.js",
+    "heroku-postbuild": "npm run build",
+    "start": "node server.js",
     "dev": "webpack-dev-server --hot --env.mode=development --open --port 5000",
     "build": "rm -rf ./dist && webpack --env.mode production",
     "test": "jest",
@@ -69,7 +70,7 @@
     "ts-loader": "^8.0.2",
     "tslint-config-prettier": "^1.18.0",
     "typescript": "^3.9.7",
-    "webpack": "^4.44.1",
+    "webpack": "^4.46.0",
     "webpack-cli": "^3.3.12",
     "webpack-dev-server": "^3.11.0",
     "webpack-merge": "^5.1.1"
