+++ src/main/components/UI/templates/ConvertTemplate/index.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { NeumorphLogo } from "../../molecules/NeumorphLogo";
+import { NeumorphLogo } from "@molecules/NeumorphLogo";
 import styles from "./index.module.css";
 
 interface ConvertTemplateDef {
