+++ src/main/routes/index.tsx
@@ -2,11 +2,11 @@ import * as React from "react";
 import { BrowserRouter as Router } from "react-router-dom";
 import { Provider } from "react-redux";
 import AppRouter from "./AppRouter";
-import "../assets/styles/_variables.css";
-import "../assets/styles/_base.css";
-import "../assets/styles/_fonts.css";
-import "../assets/styles/_commons.css";
-import store from "../store/index";
+import "@styles/_variables.css";
+import "@styles/_base.css";
+import "@styles/_fonts.css";
+import "@styles/_commons.css";
+import store from "@store/index";
 
 export const Routes = () => {
     return (
