+++ tsconfig.json
@@ -42,8 +42,18 @@
 
         /* Module Resolution Options */
         // "moduleResolution": "node",            /* Specify module resolution strategy: 'node' (Node.js) or 'classic' (TypeScript pre-1.6). */
-        // "baseUrl": "./",                       /* Base directory to resolve non-absolute module names. */
-        // "paths": {},                           /* A series of entries which re-map imports to lookup locations relative to the 'baseUrl'. */
+        "baseUrl": "./",                       /* Base directory to resolve non-absolute module names. */
+        "paths": {
+            "@atoms/*": ["src/main/components/UI/atoms/*"],
+            "@molecules/*": ["src/main/components/UI/molecules/*"],
+            "@organisms/*": ["src/main/components/UI/organisms/*"],
+            "@templates/*": ["src/main/components/UI/templates/*"],
+            "@pages/*": ["src/main/components/Pages/*"],
+            "@store/*": ["src/main/store/*"],
+            "@utils/*": ["src/main/utilities/*"],
+            "@imgs/*": ["src/main/assets/images/*"],
+            "@styles/*": ["src/main/assets/styles/*"],
+        },                           /* A series of entries which re-map imports to lookup locations relative to the 'baseUrl'. */
         // "rootDirs": [],                        /* List of root folders whose combined content represents the structure of the project at runtime. */
         // "typeRoots": [],                       /* List of folders to include type definitions from. */
         // "types": [],                           /* Type declaration files to be included in compilation. */
