+++ webpack.config.js
@@ -6,7 +6,5 @@ module.exports = (env) => {
         ? require("./webpack-build-utils/webpack.development")
         : require(`./webpack-build-utils/webpack.${env.mode}`);
 
-    // return webpackMerge({ mode: env.mode }, commonConfig, envConfig);
-    // return mergeWithCustomize({ mode: env.mode }, commonConfig, envConfig);
     return merge({ mode: env.mode }, commonConfig, envConfig);
 };
